Redesign MathBoxes for better efficiency
The redesigned `MathBox`es now perform necessary calculations lazily
to improve performance. This includes various changes:

* The new `LazyVec` type:

The content of a `MathBox` is now generic in the type of its `Boxes`
variant which is used to represent that the contents of a math box is a
list of other sub-math-boxes. In non user-facing code the type for
representing a list of math boxes is the new `LazyVec`.

A `LazyVec` is a generic enum whose value is either an `Iterator` or a
`Vec` yielding/containing items of the same type. A LazyVec's special
ability is to lazily and transparently convert itself to a vector when
indexed  access to its elements is needed but also provide an iterator
interface (through an `IntoIter` implementation) that will not create
a heap allocation for storing any of its elements. Because a LazyVec
needs to transparently mutate itself, its implementation contains a bit
of unsafe code using UnsafeCell for interior mutability. The public
interface for a LazyVec however is safe.

* Lazily calculate dimensions:

The dimensions of a math box including italic correction and top accent
attachment values are now only calculated when first accessed by some
code. To calculate values usually the math boxes sub-boxes have to be
evaluated and/or the glyph outlines have to be parsed. For correctly
positioning all elements in mathematical formulae these dimensions are
usually not required to be known for all math boxes.