Implement operator stretching using data from MATH table
For stretchy shaping mlayout now uses the harfbuzz-provided access to
the math variants and glyph-assembly data contained in the MATH table.
The selection and/or construction of the appropriate glyph for the
target stretch size requires a complicated algorithm that is vaguely
described in the official OpenType Spec. The implementation here closely
follows Fred Wang's excellently commented code in which was not merged
into harfbuzz master.

These changes make mlayout compatible with unmodified harfbuzz. However
it still requires a custom harfbuzz-sys crate because the official one
is based on an old harfbuzz version without support for the OpenType
MATH table.

Opentype spec:
https://www.microsoft.com/typography/otspec/math.htm

Fred Wang's code:
https://github.com/fred-wang/harfbuzz/blob/d776da2f678e1895ac61ed9dfba10efc6d1c82ee/src/hb-ot-shape-math.cc