Improve general rendering by improving box model
Add new advance_width field to MathBox that holds
the advance width of the box and is different
than the ink width. The ink extents are now
accessible via the extents() method of MathBox.

Extents were extended by one field named 
left_side_bearing. This is the distance from the
origin to the left ink edge.